requirements.txt
altgraph==0.17.4
anyio==4.9.0
certifi==2025.4.26
chardet==5.2.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
python-telegram-bot==22.0
pywin32-ctypes==0.2.3
reportlab==4.4.0
sniffio==1.3.1
typing_extensions==4.13.2
zipp==3.21.0
pywebview
